fix(slots): widen balance-slot typing u8→u16 so Option B's 1024 slots are reachable
The 2026-07-18 1000-delegate storm proved joins hard-stop at 256 active
slots ("no member at slot 256"): every slot/index wire field was u8, so
the 254th delegate's slot (3 + 253) wrapped/overflowed even though
MAX_CHANNEL_MEMBERS=1024 (BALANCE_SLOT_TREE_HEIGHT=10).

BALANCE-SLOT space (0..1024) widened to u16 end-to-end:
- MemberInfo.slot, SendPayload sender/recipient_index,
  RefreshPayload.member_index, InterChannelDebitPayload.sender_index,
  InterChannelTransferDescriptor.recipient_slot, ChannelMember.member_slot
- ChannelRecord/BalanceState delegate_count (+ close/withdrawal/post-close
  claim PI structs and their u16::try_from parses)
- wallet fns: build_record, build_send, build_refresh, build_burn_send,
  build_inter_channel_send, decrypt_balance, verify_snapshot,
  l1_deposit_import_digest
- wasm ABI slot params (pkg/ must be rebuilt on next deploy; JS callers
  pass plain numbers, no change needed)
- CLI: ControlledMember.slot, join arithmetic (capacity check now BEFORE
  the slot computation), send/claim arg parses

COSIGNER space (< member_count <= MAX_COSIGNERS=16) intentionally stays
u8 — bp_member_slot, MemberSignature.member_slot, sign_state — with every
narrowing cast guarded by an explicit member_count bound and documented.
New guards: build_record rejects member_count > MAX_COSIGNERS before the
u8 narrowing (258 previously truncated to 2 and validated); wallet_cosign
rejects delegate sessions (send-only model).

Digest/fixture audit: every digest encodes slots/counts as u32/u64 limbs
(IMCR, IMSB, h1, signed-small-block, deposit-import, member-set
commitment), so the widening changes no digest and invalidates no baked
fixture; circuit VKs untouched (delegate_count was already a full u32
limb Target). bincode is only used for circuit-data payloads.

Regression tests (native, no proving, ~2s):
- wallet_core::slot_capacity_tests::join_path_reaches_slot_256_and_beyond
  builds records at 256/257/300 active slots, resolves member_at(256) —
  the storm's exact failing lookup — and JSON-roundtrips slot 299.
- balance_state_validates_past_256_active_slots validates a 257-active
  BalanceState and checks H1 commits the u16 delegate_count.
- balance_state: the member_count+delegate_count > MAX capacity check is
  now reachable (dead code under u8) and is exercised.

Verified: cargo check debug+release (lib/bins/tests), wasm32 build-std
check, common/wallet_core/PI suites, release delegate_send proving suite.
(doc/benches/degree_report.rs was already broken on the base branch.)

Evidence: .claude/deploy-record.md "2026-07-18 — 1000-connection storm".

+++ tests/channel_backing_e2e.rs
@@ -48,7 +48,7 @@ const CHANNEL: u32 = 1; // base-layer user id == channel id (detail2 §A-2)
 // anvil dev account[0] (public throwaway key; NEVER a real key).
 const ANVIL0: &str = "0xac0974bec39a17e36ba4a6b4d238ff944bacb478cbed5efcae784d7bf4f2ff80";
 
-fn info(slot: u8, k: &MemberKeys) -> MemberInfo {
+fn info(slot: u16, k: &MemberKeys) -> MemberInfo {
     MemberInfo {
         slot,
         pk_g: k.pk_g(),
@@ -267,7 +267,7 @@ fn deposit_backing_gate_reconciles_and_fails_closed() {
     let members: Vec<MemberInfo> = mkeys
         .iter()
         .enumerate()
-        .map(|(i, k)| info(i as u8, k))
+        .map(|(i, k)| info(i as u16, k))
         .collect();
     let record = build_record(CHANNEL, &members, 0, 0).expect("record");
     let cts: Vec<_> = mkeys
@@ -309,7 +309,7 @@ fn deposit_backing_gate_reconciles_and_fails_closed() {
     let foreign_members: Vec<MemberInfo> = mkeys
         .iter()
         .enumerate()
-        .map(|(i, k)| info(i as u8, k))
+        .map(|(i, k)| info(i as u16, k))
         .collect();
     let foreign_record = build_record(CHANNEL + 1, &foreign_members, 0, 0).expect("foreign record");
     assert!(
